After retrieving a sample from cache as part of [upsertOneSample, patchSample, putSample, getSample, findSamples], when preparing to return the sample to the controller or publisher, test it for empty or invalid "name" attribute before trying to publish it (which is where the missing sample name issue would get thrown).
Also moved the code which splits and tests the sample name into a separate fn, instead of repeating it multiple times.

Also: pulled "sample" out of the "checkWritePermission" args cuz it was only used in the error message, when in fact we should be showing the aspect name in the error message, not the sample name.

Also: added sample/subject/aspect names to some error messages.

The changes to [api/v1/controllers/samples.js, api/v1/helpers/verbs/utils.js, utils/common.js] are "noop" changes, just formatting and comments.