Remove "paranoid" setting from db models (exc. Users & Profiles) (#1181)
* Remove "paranoid" setting from db models (exc. Users & Profiles)
Remove isDeleted attribute (exc. Users & Profiles)
Add migration to clean up the soft-deleted data and update the indices
Update tests

Note: we are keeping paranoid on for Users because we can't delete some of the soft-deleted user records due to referential integrity, where another record has a soft-deleted user associated with it as creator and/or owner. And we're keeping it on for Profiles because we're keeping it on for Users.

Changes to /api/v1/helpers/nouns/*:

pull "isDeleted" out of the list of "readOnlyFields"
Changes to /db/model/*:

remove "paranoid" setting on model
remove "isDeleted" field defn
no longer call setIsDeleted in beforeDestroy
remove "isDeleted" from list of fields in any index defns
remove "isDeleted" from list of "ignoreAttributes" when calling publishObject

Migration:

Up
- fixReferences - clear the collectorGroupId col for collectors which reference a soft-deleted collector group
- destroySoftDeleted - deletes any soft-deleted records from the db
- removeOldIndices - deletes the indices which include deletedAt/isDeleted
- removeFields - removes the deletedAt/isDeleted columns

Down
- recreateOldFields - adds back the deletedAt/isDeleted columns
- removeNewIndices - removes the new indices without deletedAt/isDeleted, if they were created

* jscs fix

* fix typo

* Update 20190603225918-no-paranoid.js